+++ package.json
@@ -18,6 +18,7 @@
     "react": "19.2.4",
     "react-dom": "19.2.4",
     "react-hook-form": "7.72.0",
+    "uuid": "13.0.0",
     "zod": "4.3.6"
   },
   "devDependencies": {
