+++ utils/index.ts
@@ -1,0 +1,3 @@
+import { v4 as uuidv4 } from 'uuid'
+
+export const generateKey = () => uuidv4()
